+++ MonsterChase/Engine/Engine.h
@@ -7,9 +7,9 @@ public:
 	~Engine();
 
 	void getPlayerInput(char * &input);
-	void * initalizeMemory(int amouunt, int sizeOfAllocation);
-	void * reinitalizeMemory(void * memoryAddress, int sizeOfAllocation);
-	void copyMemory(void * to, void * from, int sizeToCopy);
+	void * initalizeMemory(size_t amouunt, size_t sizeOfAllocation);
+	void * reinitalizeMemory(void * memoryAddress, size_t sizeOfAllocation);
+	void copyMemory(void * to, void * from, size_t sizeToCopy);
 	void freeMemory(void * memoryAddress);
 };
